+++ src/app/auth/signin/page.js
@@ -5,62 +5,13 @@ import { useState } from "react";
 import { useRouter } from "next/navigation";
 import { useDispatch } from "react-redux";
 import { setUser, setLoading, setError } from "@/app/redux/slices/authSlice";
+import SigninForm from "@/app/components/SigninForm";
 
 export default function SignIn() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [error, setErrorState] = useState(null);
-  const router = useRouter();
-  const dispatch = useDispatch();
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    dispatch(setLoading(true));
-
-    const result = await signIn("credentials", {
-      redirect: false,
-      email,
-      password,
-    });
-
-    if (result.error) {
-      dispatch(setError(result.error));
-      setErrorState(result.error);
-    } else {
-      dispatch(setUser(result.user));
-      router.push("/");
-    }
-
-    dispatch(setLoading(false));
-  };
-
   return (
-    <div>
+    <div className="flex w-full h-[60vh] justify-center">
       <h1>Sign In</h1>
-      <form onSubmit={handleSubmit}>
-        <label>
-          Email:
-          <input
-            type="email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            required
-          />
-        </label>
-        <br />
-        <label>
-          Password:
-          <input
-            type="password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            required
-          />
-        </label>
-        <br />
-        <button type="submit">Sign In</button>
-      </form>
-      {error && <p style={{ color: "red" }}>{error}</p>}
+      <SigninForm />
     </div>
   );
 }
